Preprocessor conditional compiling jazz
Added all the stuff of last commit with preprocessor if statements to
adjust stuff according to the chip selected. Would be a lot
easier/cleaner if Atmel decided to sort of standardise names of
registers and the bits in them/interrupt vector sources etc. Sigh.
Untested. Will do one day.